+++ src/rules/no-named-default.js
@@ -17,7 +17,7 @@ module.exports = {
             return;
           }
 
-          if (im.type === 'ImportSpecifier' && im.imported.name === 'default') {
+          if (im.type === 'ImportSpecifier' && (im.imported.name || im.imported.value) === 'default') {
             context.report({
               node: im.local,
               message: `Use default import syntax to import '${im.local.name}'.` });